webapp/requirements.txt
absl-py==0.4.1
astor==0.7.1
certifi==2018.8.24
chardet==3.0.4
gast==0.2.0
grpcio==1.15.0
h5py==2.8.0
idna==2.7
Keras==2.2.2
Keras-Applications==1.0.4
Keras-Preprocessing==1.0.2
Markdown==2.6.11
numpy==1.14.5
Pillow==5.2.0
protobuf==3.6.1
PyYAML==3.13
requests==2.19.1
scipy==1.1.0
six==1.11.0
tensorboard==1.10.0
tensorflow==1.10.0
termcolor==1.1.0
urllib3==1.23
Werkzeug==0.14.1
setuptools~=39.2.0
cryptography==2.2.2
Flask==0.12.1